pyproject.toml
[tool.poetry]
name = "kappo-store-tg"
version = "0.1.0"
description = ""
authors = ["altmersq <aleksandro551@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiofiles = "23.2.1"
aiogram = "3.4.1"
aiohttp = "3.9.1"
aiosignal = "1.3.1"
annotated-types = "0.6.0"
asyncpg = "0.29.0"
attrs = "23.2.0"
build = "1.2.2.post1"
cachecontrol = "0.14.0"
certifi = "2023.11.17"
cfgv = "3.4.0"
charset-normalizer = "3.4.0"
cleo = "2.1.0"
colorama = "0.4.6"
crashtest = "0.4.1"
distlib = "0.3.9"
dulwich = "0.21.7"
fastjsonschema = "2.20.0"
filelock = "3.16.1"
frozenlist = "1.4.1"
greenlet = "3.0.3"
identify = "2.6.1"
idna = "3.6"
installer = "0.7.0"
jaraco-classes = "3.4.0"
keyring = "24.3.1"
magic-filter = "1.0.12"
more-itertools = "10.5.0"
msgpack = "1.1.0"
multidict = "6.0.4"
nodeenv = "1.9.1"
packaging = "24.1"
pexpect = "4.9.0"
pkginfo = "1.11.2"
platformdirs = "4.3.6"
poetry = "1.8.4"
poetry-core = "1.9.1"
poetry-plugin-export = "1.8.0"
pre-commit = "4.0.1"
psycopg2-binary = "2.9.9"
ptyprocess = "0.7.0"
pydantic = "2.5.3"
pydantic-settings = "2.2.0"
pydantic-core = "2.14.6"
pyproject-hooks = "1.2.0"
python-dotenv = "1.0.1"
pywin32-ctypes = "0.2.3"
pyyaml = "6.0.2"
rapidfuzz = "3.10.0"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
shellingham = "1.5.4"
sqlalchemy = "2.0.27"
tomlkit = "0.13.2"
trove-classifiers = "2024.10.21.16"
typing-extensions = "4.9.0"
urllib3 = "2.2.3"
virtualenv = "20.27.0"
yarl = "1.9.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
